Finalizando o refactoring para DAO Generic
descricao opcional, vamos ver o que é isso...

+++ src/main/java/com/williansmartins/dao/Dao.java
@@ -7,9 +7,9 @@ public interface Dao<T extends Serializable> {
 
 	void insert(T entity);
 	void update(T entity);
-	void delete(Long primaryKey) throws Exception;
+	void delete(Integer primaryKey) throws Exception;
 	List<T> findAll();
-	List<T> findEspecific(Long id);
-	T findById(Long primaryKey);
+	List<T> findEspecific(Integer id);
+	T findById(Integer primaryKey);
 	
 }
